+++ Mods ChoGGi/Fix Missing Build Upgrade Icons/metadata.lua
@@ -13,7 +13,7 @@ return PlaceObj("ModDef", {
 	"code", {
 		"Code/Script.lua",
 	},
-	"description", [[If you have older mods missing the build menu icons; this is the mod for you.
+	"description", [[If you have mods missing the build menu or upgrade icons; this is the mod for you.
 
 This will validate images for all modded buildings and try to fix the image path of missing ones.