+++ Expanded Cheat Menu/Changelog.md
@@ -37,6 +37,7 @@
 - > It'll try to list files in mount points (ie: ~"Textures" or ~"Prefabs"), needs the HelperMod installed.
 - > The X close button now checks if Ctrl or Shift is being held down (see tooltip for more info).
 - > Merged View/Dump menu items in Tools.
+- > Added Dist2D len to index tables of points (context menu tooltip).
 - Console Window:
 - > Added a Clipboard button (copies log to clipboard).
 - > Text is now selectable (also removed Copy Log Text).